Refactored Reader/Writer
