+++ corefxlab.sln
@@ -138,6 +138,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Azure.Experimental",
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.IO.Pipelines.Extensions.Tests", "tests\System.IO.Pipelines.Extensions.Tests\System.IO.Pipelines.Extensions.Tests.csproj", "{64C08774-982C-4141-8F8D-2884B6FA0E4B}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Buffers.ReaderWriter", "src\System.Buffers.ReaderWriter\System.Buffers.ReaderWriter.csproj", "{C5F9D191-CA3B-4648-B8A9-62E33B4622EB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -892,6 +894,18 @@ Global
 		{64C08774-982C-4141-8F8D-2884B6FA0E4B}.Release|x64.Build.0 = Release|Any CPU
 		{64C08774-982C-4141-8F8D-2884B6FA0E4B}.Release|x86.ActiveCfg = Release|Any CPU
 		{64C08774-982C-4141-8F8D-2884B6FA0E4B}.Release|x86.Build.0 = Release|Any CPU
+		{C5F9D191-CA3B-4648-B8A9-62E33B4622EB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C5F9D191-CA3B-4648-B8A9-62E33B4622EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C5F9D191-CA3B-4648-B8A9-62E33B4622EB}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C5F9D191-CA3B-4648-B8A9-62E33B4622EB}.Debug|x64.Build.0 = Debug|Any CPU
+		{C5F9D191-CA3B-4648-B8A9-62E33B4622EB}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C5F9D191-CA3B-4648-B8A9-62E33B4622EB}.Debug|x86.Build.0 = Debug|Any CPU
+		{C5F9D191-CA3B-4648-B8A9-62E33B4622EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C5F9D191-CA3B-4648-B8A9-62E33B4622EB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C5F9D191-CA3B-4648-B8A9-62E33B4622EB}.Release|x64.ActiveCfg = Release|Any CPU
+		{C5F9D191-CA3B-4648-B8A9-62E33B4622EB}.Release|x64.Build.0 = Release|Any CPU
+		{C5F9D191-CA3B-4648-B8A9-62E33B4622EB}.Release|x86.ActiveCfg = Release|Any CPU
+		{C5F9D191-CA3B-4648-B8A9-62E33B4622EB}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -959,6 +973,7 @@ Global
 		{0BA3C0A9-64B9-445D-B085-8FFAC09D6E37} = {3079E458-D0E6-4F99-8CAB-80011D35C7DA}
 		{916370AB-B0D3-4136-850B-AA12FAB23ECD} = {4B000021-5278-4F2A-B734-DE49F55D4024}
 		{64C08774-982C-4141-8F8D-2884B6FA0E4B} = {3079E458-D0E6-4F99-8CAB-80011D35C7DA}
+		{C5F9D191-CA3B-4648-B8A9-62E33B4622EB} = {4B000021-5278-4F2A-B734-DE49F55D4024}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9DD4022C-A010-4A9B-BCC5-171566D4CB17}
