+++ src/System.Binary.Base64/Base64Decoder.cs
