+++ src/System.Binary.Base64/Base64Encoder.cs
