+++ src/System.Buffers.ReaderWriter/System/Buffers/Reader/BufferReader_search.cs
