+++ src/System.Buffers.ReaderWriter/System/Buffers/Reader/BufferReader_text.cs
