+++ src/System.Buffers.ReaderWriter/System/Buffers/Transformations/RemoveTranstomation.cs
