+++ src/System.Buffers.ReaderWriter/System/Buffers/Writer/BufferWriter_ints.cs
