+++ src/System.Buffers.ReaderWriter/System/Buffers/Writer/OutputExtensions.cs
