+++ src/System.Buffers.ReaderWriter/System/Buffers/Writer/OutputWriter.cs
