+++ src/System.Buffers.ReaderWriter/System/Buffers/Writer/TransformationFormat.cs
